Fix importing grako on Python 3.10
With Python 3.10 grako fails with:

> ImportError: cannot import name 'Mapping' from 'collections'

The python developers moved `Mapping` from `collections` to
`collections.abc` in 3.3 and decided that maintaining the
`collections.Mapping` alias was too much of a burden.  It was removed in
3.10.

There is already some compatibility code in utils.py, but it wasn't being
used at this import site - now it is.